msmp1evk: major update to remove msmp1osmevk and add graphics support
The old EVK based on the MA5D4EVK is not longer supported. Therefore we
rename the msmp1osmevk device tree blob back to msmp1evk, which is the
official name of the EVK for the SOM Som.

Furthermore we add support for the RGB panel. A separate device tree
blob "msmp1evk-rgb-panel.dtb" is provided.. The BSP with Qt5 graphics
support can be selected with "msmp1evk_qt5_defconfig".

U-Boot now supports the RGMII ethernet interface.